requirements.txt
# AI Agent Products - Dependencies

# LLM Providers (pick one)
groq>=0.4.0          # FREE tier - recommended for testing
# anthropic>=0.18.0  # Paid
# openai>=1.12.0     # Paid

# Configuration
pyyaml>=6.0

# Data Analysis Agent
pandas>=2.0.0
numpy>=1.24.0

# Optional: Excel support
openpyxl>=3.1.0

# Optional: Parquet support  
pyarrow>=14.0.0
